+++ mplapack/test/eig/common/Cget02.cpp
@@ -109,7 +109,7 @@ void Cget02(const char *trans, INTEGER const m, INTEGER const n, INTEGER const n
         if (xnorm <= zero) {
             resid = one / eps;
         } else {
-            resid = max(resid, ((bnorm / anorm) / xnorm) / eps);
+            resid = max(resid, REAL(((bnorm / anorm) / xnorm) / eps));
         }
     }
     //
