+++ mplapack/test/eig/common/Cget38.cpp
@@ -281,7 +281,7 @@ statement_10:
         //        Compare condition number for eigenvalue cluster
         //        taking its condition number into account
         //
-        v = max(two * castREAL(n) * eps * tnrm, smlnum);
+        v = max(REAL(two * castREAL(n) * eps * tnrm), smlnum);
         if (tnrm == zero) {
             v = one;
         }
@@ -295,8 +295,8 @@ statement_10:
         } else {
             tolin = v / sepin;
         }
-        tol = max(tol, smlnum / eps);
-        tolin = max(tolin, smlnum / eps);
+        tol = max(tol, REAL(smlnum / eps));
+        tolin = max(tolin, REAL(smlnum / eps));
         if (eps * (sin - tolin) > stmp + tol) {
             vmax = one / eps;
         } else if (sin - tolin > stmp + tol) {
@@ -328,8 +328,8 @@ statement_10:
         } else {
             tolin = v / sin;
         }
-        tol = max(tol, smlnum / eps);
-        tolin = max(tolin, smlnum / eps);
+        tol = max(tol, REAL(smlnum / eps));
+        tolin = max(tolin, REAL(smlnum / eps));
         if (eps * (sepin - tolin) > septmp + tol) {
             vmax = one / eps;
         } else if (sepin - tolin > septmp + tol) {
