+++ mplapack/test/eig/common/Rbdt01.cpp
@@ -160,9 +160,9 @@ void Rbdt01(INTEGER const m, INTEGER const n, INTEGER const kd, REAL *a, INTEGER
             resid = (resid / anorm) / (castREAL(n) * eps);
         } else {
             if (anorm < one) {
-                resid = (min(resid, castREAL(n) * anorm) / anorm) / (castREAL(n) * eps);
+                resid = (min(resid, REAL(castREAL(n) * anorm)) / anorm) / (castREAL(n) * eps);
             } else {
-                resid = min(resid / anorm, castREAL(n)) / (castREAL(n) * eps);
+                resid = min(REAL(resid / anorm), castREAL(n)) / (castREAL(n) * eps);
             }
         }
     }
